Consistently rename DateTimeExtensions to NSDateExtensions
