+++ Softeq.XToolkit.Common.iOS.Tests/Softeq.XToolkit.Common.iOS.Tests.csproj
@@ -107,7 +107,7 @@
     <Compile Include="Main.cs" />
     <Compile Include="AppDelegate.cs" />
     <Compile Include="Extensions\NsRangeExtensionsTests\NsRangeExtensionsTests.cs" />
-    <Compile Include="Extensions\DateTimeExtensionsTests\DateTimeExtensionsTests.cs" />
+    <Compile Include="Extensions\NSDateExtensionsTests\NSDateExtensionsTests.cs" />
   </ItemGroup>
   <ItemGroup>
     <PackageReference Include="xunit">
